+++ src/main/java/org/dti/se/finalproject1backend1/inners/models/valueobjects/warehouseproducts/WarehouseProductResponse.java
@@ -4,7 +4,6 @@ import lombok.Data;
 import org.dti.se.finalproject1backend1.inners.models.entities.Product;
 import org.dti.se.finalproject1backend1.inners.models.entities.Warehouse;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -12,5 +11,5 @@ public class WarehouseProductResponse {
     private UUID id;
     private Warehouse warehouse;
     private Product product;
-    private BigDecimal quantity;
+    private Double quantity;
 }
